#68 tidying following walk-through of PR diff
